+++ tests/tsan.rs
@@ -9,7 +9,7 @@ use scoped_threadpool::Pool;
 
 #[test]
 fn once() {
-    static mut RB: spsc::Queue<i32, usize, 4> = spsc::Queue::new();
+    static mut RB: spsc::Queue<i32, 4> = spsc::Queue::new();
 
     let rb = unsafe { &mut RB };
 
@@ -30,7 +30,7 @@ fn once() {
 
 #[test]
 fn twice() {
-    static mut RB: spsc::Queue<i32, usize, 4> = spsc::Queue::new();
+    static mut RB: spsc::Queue<i32, 5> = spsc::Queue::new();
 
     let rb = unsafe { &mut RB };
 
@@ -52,7 +52,7 @@ fn twice() {
 
 #[test]
 fn scoped() {
-    let mut rb: spsc::Queue<i32, usize, 4> = spsc::Queue::new();
+    let mut rb: spsc::Queue<i32, 5> = spsc::Queue::new();
 
     rb.enqueue(0).unwrap();
 
@@ -77,7 +77,7 @@ fn scoped() {
 fn contention() {
     const N: usize = 1024;
 
-    let mut rb: spsc::Queue<u8, usize, 4> = spsc::Queue::new();
+    let mut rb: spsc::Queue<u8, 4> = spsc::Queue::new();
 
     {
         let (mut p, mut c) = rb.split();
@@ -164,9 +164,9 @@ fn mpmc_contention() {
 fn unchecked() {
     const N: usize = 1024;
 
-    let mut rb: spsc::Queue<u8, usize, N> = spsc::Queue::new();
+    let mut rb: spsc::Queue<u8, N> = spsc::Queue::new();
 
-    for _ in 0..N / 2 {
+    for _ in 0..N / 2 - 1 {
         rb.enqueue(1).unwrap();
     }
 
@@ -175,32 +175,30 @@ fn unchecked() {
 
         Pool::new(2).scoped(move |scope| {
             scope.execute(move || {
-                for _ in 0..N / 2 {
-                    unsafe {
-                        p.enqueue_unchecked(2);
-                    }
+                for _ in 0..N / 2 - 1 {
+                    p.enqueue(2).unwrap();
                 }
             });
 
             scope.execute(move || {
                 let mut sum: usize = 0;
 
-                for _ in 0..N / 2 {
-                    sum = sum.wrapping_add(usize::from(unsafe { c.dequeue_unchecked() }));
+                for _ in 0..N / 2 - 1 {
+                    sum = sum.wrapping_add(usize::from(c.dequeue().unwrap()));
                 }
 
-                assert_eq!(sum, N / 2);
+                assert_eq!(sum, N / 2 - 1);
             });
         });
     }
 
-    assert_eq!(rb.len(), N / 2);
+    assert_eq!(rb.len(), N / 2 - 1);
 }
 
 #[test]
 fn len_properly_wraps() {
-    const N: usize = 3;
-    let mut rb: spsc::Queue<u8, usize, N> = spsc::Queue::new();
+    const N: usize = 4;
+    let mut rb: spsc::Queue<u8, N> = spsc::Queue::new();
 
     rb.enqueue(1).unwrap();
     assert_eq!(rb.len(), 1);
@@ -216,8 +214,8 @@ fn len_properly_wraps() {
 
 #[test]
 fn iterator_properly_wraps() {
-    const N: usize = 3;
-    let mut rb: spsc::Queue<u8, usize, N> = spsc::Queue::new();
+    const N: usize = 4;
+    let mut rb: spsc::Queue<u8, N> = spsc::Queue::new();
 
     rb.enqueue(1).unwrap();
     rb.dequeue();