+++ CHANGELOG.md
@@ -13,10 +13,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [breaking-change] `HistoryBuffer` is now working with const constructors and non-`Copy` data
 - [breaking-change] `HistoryBuffer::as_slice` and others now only return initialized values
 - [breaking-change] `MultiCore`/`SingleCore` is now removed from `spsc::Queue`
-- [breaking-change] `spsc::Queue::u8()` and `spsc::Queue::u16()` are now unsafe
+- [breaking-change] `spsc::Queue` is now `usize` only
+- [breaking-change] `spsc::Queue` now sacrifices one element for correctness (see issue #207), i.e. it creates an `N - 1` sized queue instead of the old that generated an size `N` queue
 - `Pool` and `MPMC` now works on `thumbv6m`
 - [breaking-change] `String` has had `utf8` related methods removed as this can be done via `str`
 - [breaking-change] No data structures implement `AsSlice` traits any more, now using `AsRef` and `AsMut`
+- `IndexMap::new()` is now a `const-fn`
 
 ## [v0.6.1] - 2021-03-02